git-prlog: Allow options to be interspersed with positional arguments
Replaces `firstPositionalArgumentIndex` with a scanner-style function
`formNextPositionalArgumentIndex(_:)` that scans through the arguments,
looking for the next thing that looks like a positional argument. This
both consolidates the logic for detecting positional arguments and
allows options to be interspersed between positional arguments.
(Previously after finding the first positional argument we'd stop
scanning forward.)